+++ lib/telemetry/logger.rb
@@ -2,6 +2,7 @@ require 'telemetry/logger/version'
 require 'telemetry/logger/defaults'
 require 'telemetry/logger/builder'
 require 'telemetry/logger/methods'
+require 'telemetry/logger/exception_handler'
 
 module Telemetry
   module Logger
@@ -9,8 +10,10 @@ module Telemetry
       include Telemetry::Logger::Defaults
       include Telemetry::Logger::Methods
       include Telemetry::Logger::Builder
+      include Telemetry::Logger::ExceptionHandler
 
       def setup(level: 'info', **opts)
+        @opts = opts
         output(**opts)
         self.log_level = level
         self