+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Telemetry::Logger
 
+## v0.2.0
+* Adding new module called `Telemetry::Logger::ExceptionHandler`
+* Adding methods for `exception` and magic for detecting `ElasticAPM` and `OpenTelemetry`
+* Adding `Upload Sarif output` step to `rubocop` GitHub action
+* Adding `SECURITY.md` file
+
 ## v0.1.1
 * Fixing issue with color option using the wrong method name
 * Removing format logger from the setup command